work

+++ backend/routers/report_router.py
@@ -1,14 +1,15 @@
-from datetime import datetime
+from datetime import timedelta
 
 import structlog
 from fastapi import APIRouter, BackgroundTasks, Depends
+from gotrue import UserResponse
 from sqlalchemy.orm import Session
 
-from backend import models
 from backend.database import get_db
-from backend.schemas.landsat.landsat_item_advanced import LandsatAdvancedItem
-from backend.schemas.structures.report_result import ReportResultError, ReportResultProcess, ReportResultSuccess
+from backend import schemas, models
+from backend.schemas.structures.report_result import *
 from backend.tasks.write_report import write_report_to_db
+from backend.utils.auth import get_current_user
 
 logger = structlog.get_logger()
 
@@ -21,26 +22,23 @@ async def generate_report(scene_id: str, background_tasks: BackgroundTasks, db: 
     report = query.first()
 
     if query.scalar():
-        db.delete(report)
+        if report.is_processed:
+            return ReportResultSuccess(
+                is_processed=report.is_processed,
+                scene_id=report.scene_id,
+                created_at=report.created_at,
+                data=LandsatAdvancedItem.model_validate(report.raw_data)
+            )
 
-    # if query.scalar():
-    #     if report.is_processed:
-    #         return ReportResultSuccess(
-    #             is_processed=report.is_processed,
-    #             scene_id=report.scene_id,
-    #             created_at=report.created_at,
-    #             data=LandsatAdvancedItem.model_validate(report.raw_data)
-    #         )
-    #
-    #     if datetime.now() - report.created_at < timedelta(minutes=10):
-    #         return ReportResultProcess(
-    #             is_processed=False,
-    #             scene_id=report.scene_id,
-    #             created_at=report.created_at
-    #         )
-    #
-    #     if datetime.now() - report.created_at > timedelta(minutes=10):
-    #         db.delete(report)
+        if datetime.now() - report.created_at < timedelta(minutes=10):
+            return ReportResultProcess(
+                is_processed=False,
+                scene_id=report.scene_id,
+                created_at=report.created_at
+            )
+
+        if datetime.now() - report.created_at > timedelta(minutes=10):
+            db.delete(report)
 
     db.add(
         models.Report(
@@ -57,7 +55,7 @@ async def generate_report(scene_id: str, background_tasks: BackgroundTasks, db: 
 
 
 @report_router.get("/get_report")
-async def get_report(scene_id: str, db: Session = Depends(get_db)):
+async def get_report(scene_id: str, db: Session = Depends(get_db)) -> ReportResult:
     query = db.query(models.Report).filter_by(scene_id=scene_id)
     report = query.first()
 
@@ -77,3 +75,13 @@ async def get_report(scene_id: str, db: Session = Depends(get_db)):
         created_at=report.created_at,
         data=LandsatAdvancedItem.model_validate_json(report.raw_data),
     )
+
+
+@report_router.get("/get_reports", response_model=list[schemas.Report])
+async def get_reports(user: UserResponse = Depends(get_current_user), db: Session = Depends(get_db)):
+    # todo: return only user reports
+    reports = (
+        db.query(models.Report).join(models.User).all()
+    )
+
+    return reports
